Update get method to only allow transfers to disk
The problem is that because of the FTP spec there is no simple way of
giving the user access to the socket as i though initially without
loosing the "226 Transfer complete." control response.

In order for the user to get the socket and resolve his processing
promise only once the transfer is complete he would need to be able to
register a listener after the send listener has been trigered.

Not only I'm not yet sure if I want to give the users acces to the
pending callbacks but it is also a bit tricky to assign a pending
callback without disturbing the current system.

Ideas are welcome.