Refactored some of the library
Changed:
  The secure sockets now logs lost server messages
  The nlist method (more details below)
  The get method (more details below)
  Added a ';' and a comment
  Added a download method

About the download method:
It is the first and hopefully only breaking change, I allowed myself a
breaking change as no one is using the project yet. The download method
is the exact copy of the old get method. It will download a remote file.

About the get method:
The get method will return a socket that will stream the remote file.
It will trigger an 'endGet' event to signal that the transfert is
complete.

About the nlist method:
The changes were aimed a not dropping a packet any more. Now it only
resolves when the proper control message is recieved and rejects
otherwise. The interface has not changed and this is not a breaking
change.